+++ modules/event_shovel_ecs_service/variables.tf
@@ -2,6 +2,12 @@ variable "alert_topic_arn" {}
 variable "alert_evaluation_periods" { default = "1" }
 variable "aws_account" {}
 variable "aws_region" {}
+
+variable "cpu" {
+  default = "512"
+  description = "CPU allocation per container. 256 is 0.25 vCPU."
+}
+
 variable "desired_count" {}
 variable "docker_image_name" {}
 variable "ecs_cluster_name" {}
@@ -9,6 +15,24 @@ variable "environment_name" {}
 variable "envvars" {}
 variable "kms_key_id" { default = "" }
 variable "log_group_name" {}
-variable "memory_reservation" { default = "500" }
+
+variable "memory_reservation" {
+  description = "How much memory to allocate to each container in MiB."
+  default = "1024"
+}
+
 variable "param_store_namespace" {}
 variable "skip" { default = "" }
+
+variable "private_subnet_ids" {
+  description = "Determines the AZ for the containers."
+  type = "list"
+}
+
+variable "vpc_cidr" {
+  description = "The VPC IP range. Used for ingress rules."
+}
+
+variable "vpc_id" {
+  description = "The VPC that this task should live in."
+}